feat(vote): modify main vote design
