+++ TwoHoSun/Screens/Home/Main/ConsiderationView.swift
@@ -30,7 +30,6 @@ enum UserVoteType {
 }
 
 struct ConsiderationView: View {
-    @State private var isVoted = false
     @State private var selectedVoteType = UserVoteType.agree
     @State private var currentVote = 0
     let viewModel: ConsiderationViewModel
@@ -87,7 +86,10 @@ extension ConsiderationView {
             TabView(selection: $currentVote) {
                 // TODO: - cell 5개로 설정해 둠
                 ForEach(0..<5) { index in
-                    voteContentCell
+                    VoteContentCell(isVoted: false,
+                                    isEnd: Bool.random(),
+                                    selectedVoteType: selectedVoteType,
+                                    currentVote: $currentVote)
                         .tag(index)
                 }
                 .rotationEffect(.degrees(-90))
@@ -99,116 +101,32 @@ extension ConsiderationView {
             .tabViewStyle(.page(indexDisplayMode: .never))
         }
     }
-
-    private var voteContentCell: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            HStack(spacing: 8) {
-                Image("imgDefaultProfile")
-                    .frame(width: 32, height: 32)
-                Text("얄루")
-                    .font(.system(size: 16, weight: .bold))
-                    .foregroundStyle(.white)
-                SpendTypeLabel(spendType: .saving, size: .large)
-            }
-            .padding(.bottom, 10)
-            Text("ACG 마운틴 플라이 할인하는데 살까?")
-                .font(.system(size: 16, weight: .bold))
-                .foregroundStyle(.white)
-            Text("텍스트 내용 100자 미만")
-                .font(.system(size: 14))
-                .foregroundStyle(.white)
-                .padding(.bottom, 8)
-            HStack(spacing: 0) {
-                Text("가격: 120,000원")
-                Text(" · ")
-                Text("2020년 3월 12일")
-            }
-            .font(.system(size: 14))
-            .foregroundStyle(Color.gray100)
-            .padding(.bottom, 10)
-            HStack {
-                voteInfoButton(label: "256명 투표", icon: "person.2.fill")
-                Spacer()
-                voteInfoButton(label: "댓글 20개", icon: "message.fill")
-            }
-            .padding(.bottom, 2)
-            RoundedRectangle(cornerRadius: 16)
-                .strokeBorder(Color.disableGray, lineWidth: 1)
-                .frame(maxWidth: .infinity)
-                .aspectRatio(1.5, contentMode: .fit)
-            VStack(spacing: 10) {
-                VoteView(isVoted: isVoted,
-                         selectedVoteType: selectedVoteType)
-                detailResultButton
-            }
-            .padding(.top, 2)
-            nextVoteButton
-                .padding(.top, 16)
-        }
-        .padding(.horizontal, 24)
-    }
-
-    private func voteInfoButton(label: String, icon: String) -> some View {
-        HStack(spacing: 2) {
-            Image(systemName: icon)
-            Text(label)
-        }
-        .font(.system(size: 14))
-        .foregroundStyle(.white)
-        .padding(.vertical, 7)
-        .padding(.horizontal, 10)
-        .background(Color.darkGray2)
-        .clipShape(.rect(cornerRadius: 34))
-    }
-
-    private var detailResultButton: some View {
-        NavigationLink {
-            DetailView(isDone: false)
-        } label: {
-            Text("상세보기")
-                .font(.system(size: 16, weight: .semibold))
-                .foregroundStyle(.white)
-                .frame(maxWidth: .infinity)
-                .frame(height: 48)
-                .background(Color.blue100)
-                .clipShape(Capsule())
-        }
-    }
-
-    @ViewBuilder
-    private var nextVoteButton: some View {
-        HStack {
-            Spacer()
-            Button {
-                withAnimation {
-                    if currentVote != 4 {
-                        currentVote += 1
-                    }
-                }
-            } label: {
-                Image("icnCaretDown")
-                    .opacity(currentVote != 4 ? 1 : 0)
-            }
-            Spacer()
-        }
-    }
-
+    
     private var createVoteButton: some View {
         NavigationLink {
             WriteView(viewModel: WriteViewModel())
         } label: {
             HStack(spacing: 2) {
                 Image(systemName: "plus")
-                    .font(.system(size: 14))
                 Text("투표만들기")
-                    .font(.system(size: 14, weight: .medium))
             }
+            .font(.system(size: 14, weight: .bold))
             .foregroundStyle(.white)
             .padding(.horizontal, 11)
             .padding(.vertical, 12)
             .background(Color.lightBlue)
             .clipShape(Capsule())
         }
+    }
+
+    private var endLabel: some View {
+        Text("종료")
+            .font(.system(size: 12, weight: .medium))
+            .foregroundStyle(.white)
+            .padding(.vertical, 4)
+            .padding(.horizontal, 5)
+            .background(Color.disableGray)
+            .clipShape(RoundedRectangle(cornerRadius: 3))
     }
 
     private func getFirstDecimalNum(_ voteRatio: Double) -> Int {
