+++ TwoHoSun/Screens/Home/Main/VoteView.swift
@@ -9,17 +9,29 @@ import SwiftUI
 
 struct VoteView: View {
     @State var isVoted = false
+    @State var isEnd = false
     @State var selectedVoteType = UserVoteType.agree
+    @State var agreeVoteRatio = 30.0
+    @State var disagreeVoteRatio = 70.0
+
+    var isAgreeVoteRatioHigher: Bool {
+        agreeVoteRatio > disagreeVoteRatio
+    }
 
     var body: some View {
         VStack(spacing: 8) {
-            if isVoted {
+            if isEnd {
+                resultVoteButton(voteType: .agree,
+                                 voteRatio: agreeVoteRatio)
+                resultVoteButton(voteType: .disagree,
+                                 voteRatio: disagreeVoteRatio)
+            } else if isVoted {
                 completedVoteButton(voteType: .agree,
                                     selectedType: selectedVoteType,
-                                    voteRatio: 50)
+                                    voteRatio: agreeVoteRatio)
                 completedVoteButton(voteType: .disagree,
                                     selectedType: selectedVoteType,
-                                    voteRatio: 50)
+                                    voteRatio: disagreeVoteRatio)
             } else {
                 incompletedVoteButton(.agree)
                 incompletedVoteButton(.disagree)
@@ -43,10 +55,25 @@ extension VoteView {
                     Image(systemName: voteType.iconImage)
                     Text(voteType.title)
                 }
-                .font(.system(size: 16))
+                .font(.system(size: 16, weight: .bold))
                 .foregroundStyle(.white)
                 .padding(.leading, 20)
             }
+        }
+        .frame(height: 48)
+    }
+
+    private func resultVoteButton(voteType: UserVoteType, voteRatio: Double) -> some View {
+        ZStack(alignment: .leading) {
+            Capsule()
+                .foregroundStyle(Color.black100)
+                .overlay(alignment: .leading) {
+                    voteBackgroundButtonView(voteType: voteType,
+                                             voteRatio: voteRatio)
+                }
+                .clipShape(Capsule())
+            voteDescriptionView(voteType: voteType,
+                                voteRatio: voteRatio)
         }
         .frame(height: 48)
     }
@@ -56,24 +83,42 @@ extension VoteView {
             Capsule()
                 .foregroundStyle(Color.black100)
                 .overlay(alignment: .leading) {
-                    let voteButtonWidth = UIScreen.main.bounds.width - 48
-                    Rectangle()
-                        .frame(width: voteButtonWidth * voteRatio * 0.01)
-                        .foregroundStyle(voteType == selectedType ? Color.lightBlue : Color.gray200)
+                    voteBackgroundButtonView(voteType: voteType,
+                                             voteRatio: voteRatio)
+                }
+                .overlay {
+                    Capsule()
+                        .strokeBorder(Color.lightBlue, lineWidth: !isEnd && voteType == selectedType ? 1 : 0)
                 }
                 .clipShape(Capsule())
-            HStack(spacing: 4) {
-                Image(systemName:voteType.iconImage)
-                Text(voteType.title)
-                Spacer()
-                Text(String(format: getFirstDecimalNum(voteRatio) == 0 ?
-                            "%.0f" : "%.1f", voteRatio) + "%")
-            }
-            .foregroundStyle(.white)
-            .font(.system(size: 16))
-            .padding(.horizontal, 20)
+            voteDescriptionView(voteType: voteType,
+                                voteRatio: voteRatio)
         }
         .frame(height: 48)
+    }
+
+    private func voteBackgroundButtonView(voteType: UserVoteType, voteRatio: Double) -> some View {
+        let voteButtonWidth = UIScreen.main.bounds.width - 48
+
+        return Rectangle()
+            .frame(width: voteButtonWidth * voteRatio * 0.01)
+            .foregroundStyle(isAgreeVoteRatioHigher && voteType == .agree ||
+                             !isAgreeVoteRatioHigher && voteType == .disagree ?
+                             Color.lightBlue : Color.gray200)
+    }
+
+    private func voteDescriptionView(voteType: UserVoteType, voteRatio: Double) -> some View {
+        HStack(spacing: 4) {
+            Image(systemName: voteType.iconImage)
+            Text(voteType.title)
+                .fontWeight(.bold)
+            Spacer()
+            Text(String(format: getFirstDecimalNum(voteRatio) == 0 ?
+                        "%.0f" : "%.1f", voteRatio) + "%")
+        }
+        .foregroundStyle(.white)
+        .font(.system(size: 16))
+        .padding(.horizontal, 20)
     }
 
     private func getFirstDecimalNum(_ voteRatio: Double) -> Int {
